Struct phylotree::distance::DistanceMatrix

pub struct DistanceMatrix<T> {
    pub size: usize,
    pub taxa: Vec<String>,
    /* private fields */
}

A phylogenetic distance matrix

Fields
size: usize

Number of taxa in the matrix

taxa: Vec<String>

Identifiers of the taxa

Implementations

impl<T> DistanceMatrix<T>
where T: Display + Debug + Float + Zero + FromStr,

pub fn new(taxa: Vec<String>, matrix: Vec<T>) -> Self

Create a new distance matrix for a certain number of sequences

pub fn new_with_size(size: usize) -> Self

Create an empty distance matrix with a given size

pub fn set_taxa(&mut self, taxa: Vec<String>) -> Result<(), MatrixError>

Set the taxa of the matrix

pub fn get(&self, id_1: &str, id_2: &str) -> Result<&T, MatrixError>

Get the distance between two sequences

pub fn set( &mut self, id_1: &str, id_2: &str, dist: T, ) -> Result<(), MatrixError>

Set the distance between two sequences

pub fn to_map(&self) -> HashMap<(String, String), T>

Get the distance matrix as a HashMap containing taxa pairs as keys and pairwise distances as values

pub fn to_phylip(&self, square: bool) -> Result<String, MatrixError>

Outputs a matrix as a phylip formatted string

pub fn to_file(&self, path: &Path, square: bool) -> Result<(), MatrixError>

Writes the matrix to a phylip file

pub fn from_phylip(phylip: &str, square: bool) -> Result<Self, ParseError<T>>

Build a distance matrix from a phylip formatted string

pub fn from_file(path: &Path, square: bool) -> Result<Self, ParseError<T>>

Reads the matrix from a phylip file