Module phylotree::tree

Build and manioulate phylogenetic trees.

This module defines the two essential structs to represent phylogenetic trees:

The Node struct that represents a node of a phylogenetic tree.

The Tree struct that holds a collection of Node objects.

Modules

draw

A module to draw phylogenetic trees A module to draw phylogenetic trees

Structs

Comparison

Struct to hold tree comparison metrics

Node

A node of the Tree

Tree

A Phylogenetic tree

Enums

NewickFormat

Newick output format

NewickParseError

Errors that can occur when parsing newick files.

NodeError

Errors that can occur when manipulating Node structs.

TreeError

Errors that can occur when reading, writing and manipulating Tree structs.

Type Aliases

Edge

A type that represents branch lengths between Node objects within phylogenetic Tree object.

NodeId

A type that represents Identifiers of Node objects within phylogenetic Tree object.