Struct phylotree::tree::Tree

pub struct Tree { /* private fields */ }

A Phylogenetic tree

Implementations

impl Tree

Base methods to add and get Node objects to and from the Tree.

pub fn new() -> Self

Create a new empty Tree object

pub fn add(&mut self, node: Node) -> NodeId

Add a new node to the tree.

pub fn add_child( &mut self, node: Node, parent: NodeId, edge: Option<Edge>, ) -> Result<NodeId, TreeError>

Add a child to one of the tree’s nodes.

Example

use phylotree::tree::{Tree,Node};

// Create the tree and add a root node
let mut tree = Tree::new();
let root_id = tree.add(Node::new());

// Add children to the root
let left = tree.add_child(Node::new(), root_id, None).unwrap();
let right = tree.add_child(Node::new(), root_id, Some(0.1)).unwrap();

assert_eq!(tree.get(&root_id).unwrap().children.len(), 2);

// The depths of child nodes are derived from the parent node
assert_eq!(tree.get(&left).unwrap().get_depth(), 1);
assert_eq!(tree.get(&right).unwrap().get_depth(), 1);

// If an edge length is specified then it is set in both child and parent
assert_eq!(tree.get(&right).unwrap().parent_edge, Some(0.1));
assert_eq!(tree.get(&root_id).unwrap().get_child_edge(&right), Some(0.1));

pub fn get(&self, id: &NodeId) -> Result<&Node, TreeError>

Get a reference to a specific Node of the tree

pub fn get_mut(&mut self, id: &NodeId) -> Result<&mut Node, TreeError>

Get a mutable reference to a specific Node of the tree

pub fn get_by_name(&self, name: &str) -> Option<&Node>

Get a reference to a node in the tree by name. Note that this does not check for name unicity, if several nodes match a name this funciton will return the first match in the tree. If you want to find all nodes matching a name in a given tree, use Tree::search_nodes.

use phylotree::tree::{Tree, Node};

let mut tree = Tree::new();
let root_idx = tree.add(Node::new_named("root"));
let child_idx = tree.add_child(Node::new_named("child"), root_idx, None).unwrap();

assert_eq!(tree.get_by_name("child"), Some(tree.get(&child_idx).unwrap()));

pub fn get_by_name_mut(&mut self, name: &str) -> Option<&mut Node>

Get a mutable reference to a node in the tree by name

pub fn search_nodes(self, cond: fn(_: &Node) -> bool) -> Vec<NodeId>

Search nodes in the tree with a closure.

use phylotree::tree::{Tree, Node};

let mut tree = Tree::new();
let root_idx = tree.add(Node::new_named("root"));
let mut indices = vec![];

for name in ["A", "B", "A"] {
    let idx = tree.add_child(Node::new_named(name), root_idx, None).unwrap();
    if name == "A" { indices.push(idx) }
}

let found = tree.search_nodes(|node| node.name == Some("A".into()));
assert_eq!(found, indices);

pub fn get_root(&self) -> Result<NodeId, TreeError>

Gets the root node. In the case of unrooted trees this node is a “virtual root” that has exactly 3 children.

pub fn get_leaves(&self) -> Vec<NodeId>

Returns a Vec containing the Node IDs of leaf nodes of the tree

use phylotree::tree::{Tree, Node};

let mut tree = Tree::new();
let root_idx = tree.add(Node::new());
let left = tree.add_child(Node::new(), root_idx, None).unwrap();
let right = tree.add_child(Node::new(), root_idx, None).unwrap();

assert_eq!(tree.get_leaves(), vec![left, right]);

pub fn get_leaf_names(&self) -> Vec<Option<String>>

Returns a Vec containing the Names of the leaf nodes of the tree

use phylotree::tree::{Tree, Node};

let mut tree = Tree::new();
let root_idx = tree.add(Node::new());
let _ = tree.add_child(Node::new_named("left"), root_idx, None).unwrap();
let _ = tree.add_child(Node::new_named("right"), root_idx, None).unwrap();

let names: Vec<_> = tree.get_leaf_names()
    .into_iter()
    .flatten()
    .collect();
assert_eq!(names, vec!["left", "right"]);

pub fn get_subtree(&self, root: &NodeId) -> Result<Vec<NodeId>, TreeError>

Gets the node ids of all the nodes in the subtree rooted at the specified node

use phylotree::tree::Tree;

let tree = Tree::from_newick("(A:0.1,B:0.2,(C:0.3,D:0.4)E:0.5)F;").unwrap();
let sub_root = tree.get_by_name("E").unwrap();
let subtree: Vec<_> = tree.get_subtree(&sub_root.id)
    .unwrap()
    .iter()
    .map(|id| tree.get(id).unwrap().name.clone())
    .flatten()
    .collect();

assert_eq!(subtree, vec!["E", "C", "D"])

pub fn get_descendants(&self, root: &NodeId) -> Result<Vec<NodeId>, TreeError>

Gets the node ids of all the nodes in the subtree rooted at the specified node, except the root

use phylotree::tree::Tree;

let tree = Tree::from_newick("(A:0.1,B:0.2,(C:0.3,D:0.4)E:0.5)F;").unwrap();
let sub_root = tree.get_by_name("E").unwrap();
let subtree: Vec<_> = tree.get_descendants(&sub_root.id)
    .unwrap()
    .iter()
    .map(|id| tree.get(id).unwrap().name.clone())
    .flatten()
    .collect();

assert_eq!(subtree, vec!["C", "D"])

pub fn get_subtree_leaves( &self, root: &NodeId, ) -> Result<Vec<NodeId>, TreeError>

Gets the node ids of all the leaves in the subtree rooted at the specified node

use phylotree::tree::Tree;

let tree = Tree::from_newick("(A:0.1,B:0.2,(C:0.3,D:0.4)E:0.5)F;").unwrap();
let sub_root = tree.get_by_name("E").unwrap();
let sub_leaves: Vec<_> = tree.get_subtree_leaves(&sub_root.id)
    .unwrap()
    .iter()
    .map(|id| tree.get(id).unwrap().name.clone())
    .flatten()
    .collect();

assert_eq!(sub_leaves, vec!["C", "D"])

impl Tree

Methods to traverse the Tree

pub fn preorder(&self, root: &NodeId) -> Result<Vec<NodeId>, TreeError>

Returns a vector containing node ids in the same order as the preorder tree traversal

use phylotree::tree::Tree;

let tree = Tree::from_newick("((A,(C,E)D)B,((H)I)G)F;").unwrap();
let preorder: Vec<_> = tree.preorder(&tree.get_root().unwrap())
    .unwrap()
    .iter()
    .map(|id| tree.get(id).unwrap().name.clone())
    .flatten()
    .collect();

assert_eq!(preorder, vec!["F", "B", "A", "D", "C", "E", "G", "I", "H"])

pub fn postorder(&self, root: &NodeId) -> Result<Vec<NodeId>, TreeError>

Returns a vector containing node ids in the same order as the postorder tree traversal

use phylotree::tree::Tree;

let tree = Tree::from_newick("((A,(C,E)D)B,((H)I)G)F;").unwrap();
let postorder: Vec<_> = tree.postorder(&tree.get_root().unwrap())
    .unwrap()
    .iter()
    .map(|id| tree.get(id).unwrap().name.clone())
    .flatten()
    .collect();

assert_eq!(postorder, vec!["A", "C", "E", "D", "B", "H", "I", "G", "F"])

pub fn inorder(&self, root: &NodeId) -> Result<Vec<NodeId>, TreeError>

Returns a vector containing node ids in the same order as the inorder tree traversal. This assumes that the tree is binary.

use phylotree::tree::Tree;

let tree = Tree::from_newick("((A,(C,E)D)B,((H)I)G)F;").unwrap();
let inorder: Vec<_> = tree.inorder(&tree.get_root().unwrap())
    .unwrap()
    .iter()
    .map(|id| tree.get(id).unwrap().name.clone())
    .flatten()
    .collect();

assert_eq!(inorder, vec!["A", "B", "C", "D", "E", "F", "H", "I", "G"])

N.B.: the resulting traversal in the example is different than the one in the wikipedia page. That is because the Node struct cannot have a right child only, so in our tree I is the left child of G.

pub fn levelorder(&self, root: &NodeId) -> Result<Vec<NodeId>, TreeError>

Returns a vector containing node ids in the same order as the levelorder tree traversal

use phylotree::tree::Tree;

let tree = Tree::from_newick("((A,(C,E)D)B,((H)I)G)F;").unwrap();
let levelorder: Vec<_> = tree.levelorder(&tree.get_root().unwrap())
    .unwrap()
    .iter()
    .map(|id| tree.get(id).unwrap().name.clone())
    .flatten()
    .collect();

assert_eq!(levelorder, vec!["F", "B", "G", "A", "D", "I", "C", "E", "H"])

impl Tree

Methods that compute characteristics and measures to describe the Tree

pub fn is_binary(&self) -> Result<bool, TreeError>

Check if the tree is Binary

pub fn is_rooted(&self) -> Result<bool, TreeError>

Checks if the tree is rooted (i.e. the root node exists and has exactly 2 children)

pub fn has_unique_tip_names(&self) -> Result<bool, TreeError>

Checks if all the tips have unique names (This check assumes that all tips have a name)

pub fn size(&self) -> usize

Returns the number of nodes in the tree

pub fn n_leaves(&self) -> usize

Returns the number of leaves in the tree

pub fn height(&self) -> Result<Edge, TreeError>

Returns the height of the tree (i.e. the number of edges or branch length sum from the root to the deepest tip)

use phylotree::tree::Tree;

let tree = Tree::from_newick("((A:0.1,B:0.2)G:0.1,(C:0.3,D:0.4)E:0.5)F;").unwrap();
assert_eq!(tree.height().unwrap(), 0.9);

let tree_no_brlen = Tree::from_newick("((A,B)G,(C,D)E)F;").unwrap();
assert_eq!(tree_no_brlen.height().unwrap(), 2.);

pub fn diameter(&self) -> Result<Edge, TreeError>

Returns the diameter of the tree (i.e. longest tip to tip distance)

use phylotree::tree::Tree;

let tree = Tree::from_newick("(A:0.1,B:0.2,(C:0.3,D:0.4)E:0.5)F;").unwrap();
assert_eq!(tree.diameter().unwrap(), 1.1);

let tree_no_brlen = Tree::from_newick("(A,B,(C,D)E)F;").unwrap();
assert_eq!(tree_no_brlen.diameter().unwrap(), 3.);

pub fn length(&self) -> Result<Edge, TreeError>

Returns the lenght of the trees (i.e. the sum of branch lenghts)

use phylotree::tree::Tree;

let tree = Tree::from_newick("(A:0.1,B:0.2,(C:0.3,D:0.4)E:0.5)F;").unwrap();
assert_eq!(tree.length().unwrap(), 1.5);

pub fn cherries(&self) -> Result<usize, TreeError>

Computes the number of cherries in a tree

pub fn colless(&self) -> Result<usize, TreeError>

Computes the Colless index for the tree. The colless index, $I_c$, measures the imbalance of a phylogenetic tree:
$$ I_c = \sum_{i \in nodes} |L_i - R_i| $$

Where $L_i$ is the number of leaves in the left subtree of node $i$ and $R_i$ the number of leaves in the right subtree of $i$.

pub fn colless_yule(&self) -> Result<f64, TreeError>

Computes the normalized colless statistic with a Yule null model:
$$ I_{yule} = \frac{I_c - n\cdot\log(n) - n(\gamma-1-\log(2))}{n} $$ Where $I_c$ is the unnormalized colless index (computed with Tree::colless()), $n$ the number of leaves and $\gamma$ the Euler constant.
(see also apTreeshape)

pub fn colless_pda(&self) -> Result<f64, TreeError>

Computes the normalized colless statistic with a PDA null model:
$$ I_{PDA} = \frac{I_c}{n^{3/2}} $$ Where $I_c$ is the unnormalized colless index (computed with Tree::colless()) and $n$ the number of leaves.
(see also apTreeshape)

pub fn sackin(&self) -> Result<usize, TreeError>

Computes the Sackin index. The Sackin index, $I_s$, is computed by taking the sum over all internal nodes of the number of leaves descending from that node. A smaller Sackin index means a more balanced tree.

pub fn sackin_yule(&self) -> Result<f64, TreeError>

Computes the normalized Sackin index with a Yule null model: $$ I_{yule} = \frac{I_s - 2n\cdot \sum_{j=2}^n \frac{1}{j}}{n} $$ With $I_s$ the unnormalized Sackin index (computed with Tree::sackin()) and $n$ the number of leaves in the tree.
(see also apTreeshape)

pub fn sackin_pda(&self) -> Result<f64, TreeError>

Computes the normalized sackin statistic with a PDA null model: $$ I_{PDA} = \frac{I_s}{n^{3/2}} $$ With $I_s$ the unnormalized Sackin index (computed with Tree::sackin()) and $n$ the number of leaves in the tree.
(see also apTreeshape)

impl Tree

Methods that compute edge bipartitions and compare Tree objects with each other.

pub fn get_partitions(&self) -> Result<HashSet<FixedBitSet>, TreeError>

Get all partitions of a tree

pub fn reset_bipartition_cache(&mut self)

Resets the caches used when computing bipartitions (i.e. with Tree::compare_topologies()). You should call this if you have computed bipartitions in the tree and then changed the tree.

pub fn robinson_foulds(&self, other: &Self) -> Result<usize, TreeError>

Computes the Robinson Foulds distance (Robinson & Foulds, 1981) between two trees. The RF distance is defined as the number of unique bipartitions for each tree: $$ RF = |A\cup B| - |A\cap B| $$ Where $A$ and $B$ are the sets of bipartitions of the first and second trees.
See also Tree::compare_topologies()

pub fn robinson_foulds_norm(&self, other: &Self) -> Result<f64, TreeError>

Computes the normalized Robinson Foulds distance between two trees (Robinson & Foulds, 1981). The RF distance is normalized by the maximum possible RF distance for both trees (i.e the number of bipartitions in both trees) so that the resulting distance is contained within [0, 1]:
$$ RF_{norm} = \frac{RF}{|A| + |B|} $$ Where $A$ and $B$ are the sets of bipartitions of the first and second trees.
See also Tree::compare_topologies()

pub fn weighted_robinson_foulds(&self, other: &Self) -> Result<f64, TreeError>

Computes the weighted Robinson Foulds distance between two trees (Robinson & Foulds, 1979). This distance is equal to the absolute difference of branch lengths for matched bipartitions between the two trees, plus branch lenghts for unique bipartitions:
$$ RF_{weighted} = \sum_{e \in A\cap B} |d_{(e,A)} - d_{(e,B)}| + \sum_{e \in A\setminus B}d_{(e,A)} + \sum_{e \in B\setminus A}d_{(e,B)} $$ Where $A$ and $B$ are the sets of bipartitions of the first and second trees, and $d_{(e,A)}$ the branch length of bipartition $e$ in the first tree ($A$).
See also Tree::compare_topologies()

pub fn khuner_felsenstein(&self, other: &Self) -> Result<f64, TreeError>

Computes the khuner felsenstein branch score between two trees, (Khuner & Felsenstein, 1994). The distance is computed by taking the squared difference of branch lengths for matched bipartitions between the two trees, plus squared branch lenghts for unique bipartitions. The branch score is then derived by taking the square root of that total sum:
$$ KF = \sqrt{ \sum_{e \in A\cap B} (d_{(e,A)} - d_{(e,B)})^2 + \sum_{e \in A\setminus B}d_{(e,A)}^2 + \sum_{e \in B\setminus A}d_{(e,B)}^2 } $$ See also Tree::compare_topologies()

pub fn compare_topologies(&self, other: &Self) -> Result<Comparison, TreeError>

Compute several the RF metric, the weighted and normalized RF metrics and the KF branch score in one pass. This is more efficient than calling the different functions separately.

use phylotree::tree::Tree;

let tree1 = Tree::from_newick("(A:0.1,B:0.2,(C:0.3,D:0.4)E:0.5)F;").unwrap();
let tree2 = Tree::from_newick("(A:0.1,D:0.2,(C:0.3,B:0.4)E:0.5)F;").unwrap();

let rf = tree1.robinson_foulds(&tree2).unwrap() as f64;
let norm_rf = tree1.robinson_foulds_norm(&tree2).unwrap();
let weighted_rf = tree1.weighted_robinson_foulds(&tree2).unwrap();
let branch_score = tree1.khuner_felsenstein(&tree2).unwrap();

let comparison = tree1.compare_topologies(&tree2).unwrap();

assert_eq!(rf, comparison.rf);
assert_eq!(norm_rf, comparison.norm_rf);
assert_eq!(weighted_rf, comparison.weighted_rf);
assert_eq!(branch_score, comparison.branch_score);

pub fn compare_branch_lengths( &self, other: &Self, include_tips: bool, ) -> Result<(Vec<Edge>, Vec<Edge>, Vec<(Edge, Edge)>), TreeError>

Compare sets of branches between 2 trees. This will return 3 sets of branch lengths:

branches exclusive to the self tree

branches exclusive to the other tree

2-uples for common branches: (self_len, other_len)

You can specify wether to include terminal branches in this calculation or not with the include_terminal parameter.

Example

use phylotree::tree::Tree;

// Ref Tree:     
//               0.3
//         0.2 +----- A
//   0.1 +-----| 0.4      
// +-----|     +----- B
// |     | 0.5            
// |     +----- C      
// |       0.7            
// | 0.6 +----- D      
// +-----| 0.8            
//       +----- E       

// Compared Tree:
//               0.3
//         0.2 +----- A
//   0.1 +-----| 0.4     
// +-----|     +----- C
// |     | 0.5            
// |     +----- B      
// |       7.0            
// | 0.6 +----- D      
// +-----| 0.8            
//       +----- E
//
// We just switched the B and C labels, so we should have 1
// exlusive branch per tree. The compared tree also has one branch
// that is 7.0 instead of 0.7.

let reftree = Tree::from_newick("(((A:0.3,B:0.4):0.2,C:0.5):0.1,(D:0.7,E:0.8):0.6);").unwrap();
let cmptree = Tree::from_newick("(((A:0.3,C:0.4):0.2,B:0.5):0.1,(D:7.0,E:0.8):0.6);").unwrap();
 
// Get branch comparison including terminal branches
let (refset, cmpset, common) = reftree.compare_branch_lengths(&cmptree, true).unwrap();

assert_eq!(refset, vec![0.2]); // AB,CDE exclusive to reftree
assert_eq!(cmpset, vec![0.2]); // AC,BDE exclusive to cmptree

let expected_common = [
   (0.3, 0.3), // A,...
   (0.4, 0.4), // B,...
   (0.5, 0.5), // C,...
   (0.7, 0.7), // ABC,DE
   (0.7, 7.0), // D,...
   (0.8, 0.8), // E,...
];

assert_eq!(expected_common.len(), common.len());
for val in common {
   let mut found = false;
   for exp in expected_common.iter() {
       if val.0 - exp.0 < f64::EPSILON && val.1 - exp.1 < f64::EPSILON {
           found = true;
       }
   }
   assert!(found);
}

impl Tree

Methods to find paths in a Tree as well as measure distances between Node objects.

pub fn get_path_from_root( &self, node: &NodeId, ) -> Result<Vec<NodeId>, TreeError>

Returns the path from the node to the root

use phylotree::tree::Tree;

let tree = Tree::from_newick("((A,(C,E)D)B,((H)I)G)F;").unwrap();
let path: Vec<_> = tree.get_path_from_root(&5)
    .unwrap()
    .iter()
    .map(|id| tree.get(id).unwrap().name.clone())
    .flatten()
    .collect();

assert_eq!(path, vec!["F", "B", "D", "E"])

pub fn get_common_ancestor( &self, source: &NodeId, target: &NodeId, ) -> Result<usize, TreeError>

Gets the most recent common ancestor between two tree nodes

use phylotree::tree::Tree;

let tree = Tree::from_newick("((A,(C,E)D)B,((H)I)G)F;").unwrap();
let ancestor = tree.get_common_ancestor(
    &tree.get_by_name("A").unwrap().id,
    &tree.get_by_name("D").unwrap().id,
).unwrap();

assert_eq!(tree.get(&ancestor).unwrap().name, Some("B".to_owned()))

pub fn get_distance( &self, source: &NodeId, target: &NodeId, ) -> Result<(Option<f64>, usize), TreeError>

Gets the distance between 2 nodes, returns the sum of branch lengths (if all branches in the path have lengths) and the number of edges in the path.

use phylotree::tree::Tree;

let tree = Tree::from_newick("((A,(C,E)D)B,((H)I)G)F;").unwrap();
let (sum_edge_lengths, num_edges) = tree.get_distance(
    &tree.get_by_name("A").unwrap().id,
    &tree.get_by_name("I").unwrap().id,
).unwrap();

assert_eq!(num_edges, 4);
assert!(sum_edge_lengths.is_none());

pub fn distance_matrix_recursive( &self, ) -> Result<DistanceMatrix<Edge>, TreeError>

Computes the distance matrix of the tree, implemented in a recursive manner.

use phylotree::tree::Tree;

let tree = Tree::from_newick("((T3:0.2,T1:0.2):0.3,(T2:0.4,T0:0.5):0.6);").unwrap();
let matrix = tree.distance_matrix_recursive().unwrap();

let phylip="\
4
T0    0  1.6  0.9  1.6
T1    1.6  0  1.5  0.4
T2    0.9  1.5  0  1.5
T3    1.6  0.4  1.5  0
";

assert_eq!(phylip, matrix.to_phylip(true).unwrap())

pub fn distance_matrix(&self) -> Result<DistanceMatrix<f64>, TreeError>

Computes the distance matrix of the tree.

use phylotree::tree::Tree;

let tree = Tree::from_newick("((T3:0.2,T1:0.2):0.3,(T2:0.4,T0:0.5):0.6);").unwrap();
let matrix = tree.distance_matrix().unwrap();

let phylip="\
4
T0    0  1.6  0.9  1.6
T1    1.6  0  1.5  0.4
T2    0.9  1.5  0  1.5
T3    1.6  0.4  1.5  0
";

assert_eq!(phylip, matrix.to_phylip(true).unwrap())

impl Tree

Methods to manipulate and alter the Tree object.

pub fn prune(&mut self, root: &NodeId) -> Result<(), TreeError>

Prune the subtree starting at a given root node.

Example

use phylotree::tree::Tree;

let mut tree = Tree::from_newick("((A,(C,E)D)B,((H)I)G)F;").unwrap();
let root_idx = tree.get_by_name("G").unwrap().id;

tree.prune(&root_idx);

assert_eq!(tree.to_newick().unwrap(), "((A,(C,E)D)B)F;")

pub fn compress(&mut self) -> Result<(), TreeError>

Compress the tree (i.e. remove nodes with exactly 1 parent and 1 child and fuse branches together)

use phylotree::tree::Tree;

let mut tree = Tree::from_newick("((A,(C,E)D)B,((H)I)G)F;").unwrap();
// Compress F->G->I->H to F->H
tree.compress().unwrap();

assert_eq!(tree.to_newick().unwrap(), "((A,(C,E)D)B,H)F;")

pub fn rescale(&mut self, factor: f64)

Rescale the branch lenghts of the tree

use phylotree::tree::Tree;

let mut tree = Tree::from_newick("(A:0.1,B:0.2,(C:0.3,D:0.4)E:0.5)F;").unwrap();
// Double all branch lengths
tree.rescale(2.0);

assert_eq!(
    tree.to_newick().unwrap(),
    "(A:0.2,B:0.4,(C:0.6,D:0.8)E:1)F;"
)

pub fn resolve(&mut self) -> Result<(), TreeError>

Randomly resolve multifurcations to binarize the tree

use phylotree::tree::Tree;

let mut tree = Tree::from_newick("((A:0.1,B:0.2):0.3, (C:0.1,D:0.2,E:0.4)F:0.5)G;").unwrap();
assert!(!tree.is_binary().unwrap());

tree.resolve();
assert!(tree.is_binary().unwrap());

pub fn ladderize(&mut self) -> Result<(), TreeError>

Sort children of a node by number of descendants

use phylotree::tree::Tree;

let mut tree = Tree::from_newick("(A,(((D,(E,F)),C),B));").unwrap();
tree.ladderize();

assert_eq!("(A,(B,(C,(D,(E,F)))));", tree.to_newick().unwrap());

impl Tree

Methods to read and write Tree objects to and from files or String objects.

pub fn to_newick(&self) -> Result<String, TreeError>

Writes the tree as a newick formatted string

Example

use phylotree::tree::Tree;

let newick = "(A:0.1,B:0.2,(C:0.3,D:0.4)E:0.5)F:0.6;";
let tree = Tree::from_newick(newick).unwrap();

assert_eq!(tree.to_newick().unwrap(), newick);

pub fn to_formatted_newick( &self, format: NewickFormat, ) -> Result<String, TreeError>

Writes the tree as a newick formatted string with a specified output format from NewickFormat.

Example

use phylotree::tree::{Tree, NewickFormat};

let newick = "(A:0.1,B:0.2,(C:0.3,D:0.4)E:0.5)F:0.6;";
let tree = Tree::from_newick(newick).unwrap();

assert_eq!(tree.to_formatted_newick(NewickFormat::Topology).unwrap(), "(,,(,));");
assert_eq!(
    tree.to_formatted_newick(NewickFormat::OnlyNames).unwrap(),
    "(A,B,(C,D)E)F;"
);
assert_eq!(
    tree.to_formatted_newick(NewickFormat::InternalLengthsLeafNames).unwrap(),
    "(A,B,(C,D):0.5):0.6;"
);

pub fn from_newick(newick: &str) -> Result<Self, NewickParseError>

Read a newick formatted string and build a Tree struct from it.

Example

use phylotree::tree::Tree;

let newick = "(A:0.1,B:0.2,(C:0.3,D:0.4)E:0.5)F;";
let tree = Tree::from_newick(newick).unwrap();

assert_eq!(tree.size(), 6);
assert_eq!(tree.n_leaves(), 4);
assert_eq!(tree.is_rooted().unwrap(), false);

pub fn to_file(&self, path: &Path) -> Result<(), TreeError>

Writes the tree to a newick file

pub fn from_file(path: &Path) -> Result<Self, NewickParseError>

Creates a tree from a newick file

pub fn print_debug(&self) -> Result<(), TreeError>

Print a debug view of the tree to the console

pub fn print(&self) -> Result<(), TreeError>

Print the tree to the console